+++ pkg/transforms/deployment.go
@@ -11,12 +11,12 @@ func transformDeployment(resource *v1.Deployment) Node {
 
 	// Extract the properties specific to this type
 	deployment.Properties["kind"] = "Deployment"
-	deployment.Properties["available"] = resource.Status.AvailableReplicas
-	deployment.Properties["current"] = resource.Status.Replicas
-	deployment.Properties["ready"] = resource.Status.ReadyReplicas
-	deployment.Properties["desired"] = int32(0)
+	deployment.Properties["available"] = int64(resource.Status.AvailableReplicas)
+	deployment.Properties["current"] = int64(resource.Status.Replicas)
+	deployment.Properties["ready"] = int64(resource.Status.ReadyReplicas)
+	deployment.Properties["desired"] = int64(0)
 	if resource.Spec.Replicas != nil {
-		deployment.Properties["desired"] = *(resource.Spec.Replicas)
+		deployment.Properties["desired"] = int64(*resource.Spec.Replicas)
 	}
 
 	return deployment
