+++ pkg/transforms/job.go
@@ -11,14 +11,14 @@ func transformJob(resource *v1.Job) Node {
 
 	// Extract the properties specific to this type
 	job.Properties["kind"] = "Job"
-	job.Properties["successful"] = resource.Status.Succeeded
-	job.Properties["completions"] = int32(0)
+	job.Properties["successful"] = int64(resource.Status.Succeeded)
+	job.Properties["completions"] = int64(0)
 	if resource.Spec.Completions != nil {
-		job.Properties["completions"] = *resource.Spec.Completions
+		job.Properties["completions"] = int64(*resource.Spec.Completions)
 	}
-	job.Properties["parallelism"] = int32(0)
+	job.Properties["parallelism"] = int64(0)
 	if resource.Spec.Completions != nil {
-		job.Properties["parallelism"] = *resource.Spec.Parallelism
+		job.Properties["parallelism"] = int64(*resource.Spec.Parallelism)
 	}
 
 	return job
